+++ rails/app/views/products/index.html.slim
@@ -15,7 +15,7 @@
         = f.submit class: "ui teal button",id: 'search'
 
   - if @products.present?
-    .ui.three.stackable.cards.segment.scroll#sidebar data-controller='transition sidebar'
+    .ui.three.stackable.cards.segment.scroll#sidebar data-controller='transition sidebar scroll'
       .ui.right.close.rail
         .ui.sticky
           .ui.labeled.icon.teal.button#top-button data-action='click->sidebar#top'
